requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
agno 
ag-ui-protocol
